Refactor data in LibriSpeech recipe [part 2]
