+++ egs/librispeech/asr/simple_v1/mmi_att_transformer_train.py
@@ -27,7 +27,7 @@ from snowfall.common import describe, str2bool
 from snowfall.common import load_checkpoint, save_checkpoint
 from snowfall.common import save_training_info
 from snowfall.common import setup_logger
-from snowfall.data.librispeech import LibriSpeechDataModule
+from snowfall.data.librispeech import LibriSpeechAsrDataModule, LibriSpeechDataModule
 from snowfall.models import AcousticModel
 from snowfall.models.conformer import Conformer
 from snowfall.models.transformer import Noam, Transformer
@@ -404,7 +404,7 @@ def get_parser():
 
 def main():
     parser = get_parser()
-    LibriSpeechDataModule.add_arguments(parser)
+    LibriSpeechAsrDataModule.add_arguments(parser)
     args = parser.parse_args()
 
     model_type = args.model_type
@@ -448,7 +448,7 @@ def main():
     P.scores = torch.zeros_like(P.scores)
     P = P.to(device)
 
-    librispeech = LibriSpeechDataModule(args)
+    librispeech = LibriSpeechAsrDataModule(args)
     train_dl = librispeech.train_dataloaders()
     valid_dl = librispeech.valid_dataloaders()
