+++ snowfall/data/librispeech.py
@@ -1,75 +1,32 @@
-from typing import List
+import argparse
+
+from functools import lru_cache
 
 import logging
+from typing import List
 
-import argparse
-from torch.utils.data import DataLoader
-
-from lhotse import Fbank, FbankConfig, load_manifest
-from lhotse.dataset import BucketingSampler, CutConcatenate, CutMix, K2SpeechRecognitionDataset, SingleCutSampler, \
-    SpecAugment
-from lhotse.dataset.input_strategies import OnTheFlyFeatures
+from lhotse import CutSet, load_manifest
 from snowfall.common import str2bool
-from snowfall.data.datamodule import DataModule
+from snowfall.data.asr_datamodule import AsrDataModule
 
 
-class LibriSpeechDataModule(DataModule):
+class LibriSpeechAsrDataModule(AsrDataModule):
+    """
+    LibriSpeech ASR data module. Can be used for 100h subset (``--full-libri false``) or full 960h set.
+    The train and valid cuts for standard Libri splits are concatenated into a single CutSet/DataLoader.
+    """
+
     def add_arguments(self, parser: argparse.ArgumentParser):
-        group = parser.add_argument_group(
-            title='Data related options',
-            description='These options are used for the preparation of PyTorch DataLoaders '
-                        'from Lhotse CutSet\'s -- they control the effective batch sizes, '
-                        'sampling strategies, applied data augmentations, etc.'
-        )
-        group.add_argument(
-            '--max-duration',
-            type=int,
-            default=500.0,
-            help="Maximum pooled recordings duration (seconds) in a single batch.")
-        group.add_argument(
-            '--bucketing_sampler',
-            type=str2bool,
-            default=False,
-            help='When enabled, the batches will come from buckets of '
-                 'similar duration (saves padding frames).')
-        group.add_argument(
-            '--num-buckets',
-            type=int,
-            default=30,
-            help='The number of buckets for the BucketingSampler'
-                 '(you might want to increase it for larger datasets).')
-        group.add_argument(
-            '--concatenate-cuts',
-            type=str2bool,
-            default=True,
-            help='When enabled, utterances (cuts) will be concatenated '
-                 'to minimize the amount of padding.')
-        group.add_argument(
-            '--duration-factor',
-            type=float,
-            default=1.0,
-            help='Determines the maximum duration of a concatenated cut '
-                 'relative to the duration of the longest cut in a batch.')
-        group.add_argument(
-            '--gap',
-            type=float,
-            default=1.0,
-            help='The amount of padding (in seconds) inserted between concatenated cuts. '
-                 'This padding is filled with noise when noise augmentation is used.')
+        super().add_arguments(parser)
+        group = parser.add_argument_group(title='LibriSpeech specific options')
         group.add_argument(
             '--full-libri',
             type=str2bool,
-            default=False,
+            default=True,
             help='When enabled, use 960h LibriSpeech.')
-        group.add_argument(
-            '--on-the-fly-feats',
-            type=str2bool,
-            default=False,
-            help='When enabled, use on-the-fly cut mixing and feature extraction. '
-                 'Will drop existing precomputed feature manifests if available.'
-        )
 
-    def train_dataloaders(self) -> DataLoader:
+    @lru_cache()
+    def train_cuts(self) -> CutSet:
         logging.info("About to get train cuts")
         cuts_train = load_manifest(self.args.feature_dir / 'cuts_train-clean-100.json.gz')
         if self.args.full_libri:
@@ -78,114 +35,22 @@ class LibriSpeechDataModule(DataModule):
                     load_manifest(self.args.feature_dir / 'cuts_train-clean-360.json.gz') +
                     load_manifest(self.args.feature_dir / 'cuts_train-other-500.json.gz')
             )
+        return cuts_train
 
-        logging.info("About to get Musan cuts")
-        cuts_musan = load_manifest(self.args.feature_dir / 'cuts_musan.json.gz')
-
-        logging.info("About to create train dataset")
-        transforms = [CutMix(cuts=cuts_musan, prob=0.5, snr=(10, 20))]
-        if self.args.concatenate_cuts:
-            logging.info(f'Using cut concatenation with duration factor '
-                         f'{self.args.duration_factor} and gap {self.args.gap}.')
-            # Cut concatenation should be the first transform in the list,
-            # so that if we e.g. mix noise in, it will fill the gaps between different utterances.
-            transforms = [
-                CutConcatenate(
-                    duration_factor=self.args.duration_factor,
-                    gap=self.args.gap
-                )
-             ] + transforms
-
-        input_transforms = [
-            SpecAugment(num_frame_masks=2, features_mask_size=27, num_feature_masks=2, frames_mask_size=100)
-        ]
-
-        train = K2SpeechRecognitionDataset(
-            cuts_train,
-            cut_transforms=transforms,
-            input_transforms=input_transforms
-        )
-
-        if self.args.on_the_fly_feats:
-            # NOTE: the PerturbSpeed transform should be added only if we remove it from data prep stage.
-            # # Add on-the-fly speed perturbation; since originally it would have increased epoch
-            # # size by 3, we will apply prob 2/3 and use 3x more epochs.
-            # # Speed perturbation probably should come first before concatenation,
-            # # but in principle the transforms order doesn't have to be strict (e.g. could be randomized)
-            # transforms = [PerturbSpeed(factors=[0.9, 1.1], p=2 / 3)] + transforms
-            # Drop feats to be on the safe side.
-            cuts_train = cuts_train.drop_features()
-            train = K2SpeechRecognitionDataset(
-                cuts=cuts_train,
-                cut_transforms=transforms,
-                input_strategy=OnTheFlyFeatures(Fbank(FbankConfig(num_mel_bins=80))),
-                input_transforms=input_transforms
-            )
-
-        if self.args.bucketing_sampler:
-            logging.info('Using BucketingSampler.')
-            train_sampler = BucketingSampler(
-                cuts_train,
-                max_duration=self.args.max_duration,
-                shuffle=True,
-                num_buckets=self.args.num_buckets
-            )
-        else:
-            logging.info('Using SingleCutSampler.')
-            train_sampler = SingleCutSampler(
-                cuts_train,
-                max_duration=self.args.max_duration,
-                shuffle=True,
-            )
-        logging.info("About to create train dataloader")
-        train_dl = DataLoader(
-            train,
-            sampler=train_sampler,
-            batch_size=None,
-            num_workers=4,
-        )
-        return train_dl
-
-    def valid_dataloaders(self) -> DataLoader:
+    @lru_cache()
+    def valid_cuts(self) -> CutSet:
         logging.info("About to get dev cuts")
-        cuts_dev = (
+        cuts_valid = (
                 load_manifest(self.feature_dir / 'cuts_dev-clean.json.gz') +
                 load_manifest(self.feature_dir / 'cuts_dev-other.json.gz')
         )
+        return cuts_valid
 
-        logging.info("About to create dev dataset")
-        if self.args.on_the_fly_feats:
-            cuts_dev = cuts_dev.drop_features()
-            validate = K2SpeechRecognitionDataset(
-                cuts_dev.drop_features(),
-                input_strategy=OnTheFlyFeatures(Fbank(FbankConfig(num_mel_bins=80)))
-            )
-        else:
-            validate = K2SpeechRecognitionDataset(cuts_dev)
-        valid_sampler = SingleCutSampler(
-            cuts_dev,
-            max_duration=self.args.max_duration,
-        )
-        logging.info("About to create dev dataloader")
-        valid_dl = DataLoader(
-            validate,
-            sampler=valid_sampler,
-            batch_size=None,
-            num_workers=1
-        )
-        return valid_dl
-
-    def test_dataloaders(self) -> List[DataLoader]:
+    @lru_cache()
+    def test_cuts(self) -> List[CutSet]:
         test_sets = ['test-clean', 'test-other']
+        cuts = []
         for test_set in test_sets:
             logging.debug("About to get test cuts")
-            cuts_test = load_manifest(self.args.feature_dir / f'cuts_{test_set}.json.gz')
-            logging.debug("About to create test dataset")
-            test = K2SpeechRecognitionDataset(
-                cuts_test,
-                input_strategy=OnTheFlyFeatures(Fbank(FbankConfig(num_mel_bins=80)))
-            )
-            sampler = SingleCutSampler(cuts_test, max_duration=self.args.max_duration)
-            logging.debug("About to create test dataloader")
-            test_dl = DataLoader(test, batch_size=None, sampler=sampler, num_workers=1)
-            yield test_dl
+            cuts.append(load_manifest(self.args.feature_dir / f'cuts_{test_set}.json.gz'))
+        return cuts
